+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/preview-middleware
 
+## 0.13.2
+
+### Patch Changes
+
+-   Updated dependencies [87c942e5]
+    -   @sap-ux/project-access@1.19.4
+    -   @sap-ux/adp-tooling@0.9.6
+
+## 0.13.1
+
+### Patch Changes
+
+-   25ecdb90: Fix missing title and description on launchpad tile
+
 ## 0.13.0
 
 ### Minor Changes